# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Web Appliance Dashboard API
  description: RESTful API für die Verwaltung von Web-Appliances, Services und SSH-Verbindungen
  version: 1.0.3
  contact:
    name: Alf Lewerken
    email: support@example.com
    url: https://github.com/alflewerken/web-appliance-dashboard
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.appliance-dashboard.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    Service:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, failed]
        running:
          type: boolean
        enabled:
          type: boolean
        uptime:
          type: string
        memory:
          type: string
        cpu:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /services/{appliance_id}/status:
    get:
      tags:
        - Service Control
      summary: Service-Status abrufen
      parameters:
        - name: appliance_id
          in: path
          required: true
          schema:
            type: integer
        - name: service
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Service-Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
  /services/{appliance_id}/start:
    post:
      tags:
        - Service Control
      summary: Service starten
      parameters:
        - name: appliance_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - service
              properties:
                service:
                  type: string
      responses:
        '200':
          description: Service gestartet
  /services/{appliance_id}/stop:
    post:
      tags:
        - Service Control
      summary: Service stoppen
      parameters:
        - name: appliance_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - service
              properties:
                service:
                  type: string
      responses:
        '200':
          description: Service gestoppt
...